refactor(review): act on the cleanup review of the context-pack fixes
Four findings, plus one proven by running.

The two static workflow tests read bare relative paths, so they passed only from
the repo root: `cd scripts/review && node --test build-context-pack.test.mjs`
failed both with ENOENT. CI passed because GitHub Actions happens to run from
the root. They are also now in the modules that actually parse the flags they
assert -- `--base`/`--body-file` with build-review-input, `--base` with
rubric-eval -- so a contributor renaming an argument meets the test that ties it
to a workflow, and both resolve their path from the module the way every other
test in this repo does.

`matches()` was called twice per expected finding with identical arguments.
Threading the PR body through this diff therefore required editing both sites,
and missing one would have been silent: `claimed` would have been built without
the body exclusion, the EXTRA list would have shrunk, and recall would have
printed the same number. Called once now.

The tokenize-and-stem expression appeared three times, differing only in the
minimum word length. The relationship that makes the body exclusion sound -- the
body must be tokenized at least as permissively as the expected terms -- was
held by nothing but two lines being adjacent. One helper, three call sites
stating their own minimum, quantifiers verified identical to what they replaced.

`typeof body === 'string' && body.trim() !== ''` gated both the reservation and
the spend; `bodyReserve > 0` holds exactly when it does. Two copies of one rule
that must not drift.

`ranked` allocated a second full-length array and walked every candidate -- about
14,400 on a large PR -- when the consumer reads at most 40. Dedup folded into
that loop; ordering, and so the "keeps its highest-ranked key" property, is
unchanged, and its test still fails when the ranking is removed.

Not taken: `rank()` inside the comparator, measured at 0.65 ms against 300 git
subprocesses in the same call, so under 1% of the function and invisible against
the model call. Deferred: `showAtRef` spawns `git show` for every changed file
even after the budget is spent; pre-existing, and this diff only makes it
marginally more reachable.
